feat(node): add behaviours with TCP transport
We need TCP instead of UDP/QUIC currently, as it seems that the AutoNAT
implemetation combined with QUIC does not properly detect the NAT
status. This means we will use TCP for now, which is more properly
tested with libp2p.

We use AutoNAT and circuit relay behaviour to let nodes behind a NAT
join the network too.