ci: remove merge workflow run on push to main
Since we are trying to only run unit tests on the merge workflow, these tests should be
deterministic and should not require other runs for extra verification.

This also disables the same workflow running on the merge queue to save the test suite running
twice.

Also added notifications for failures for the nightly full test run.