feat: enable otel tracing for the node
Enables Open Telemetry tracing for the node so we can start getting traces in Opensearch. The
configuration is almost identical to the old node. As with the old setup, `otlp` is a feature that
needs to be enabled, and it will also be toggled by use of the `OTEL_EXPORTER_OTLP_ENDPOINT`
environment variable.

I also made a change so that logging for the service always goes to stdout, even when a file is
specified for logs to go to. This is useful for quick debugging when the node is running as a
service, which it does in the remote setup. You can use `journalctl` to tail or dump out the logs,
and that would apply even if logs to file weren't enabled.

Traces are only submitted to the telemetry collector service when spans are used, so to get things
started, one span was added for swarm events; more can be added on an ad-hoc basis when needed.