chore: make the `--root-dir` argument optional
The usage of this argument was a bit confusing, because it was specified as optional, but then the
code asserted that a value had to be provided. This time it's changed to actually be optional and
the default value is indicated by the doc comments.

I took the opportunity to elaborate the documentation for the other arguments.