chore: extract sn_domain crate
Some of the functionality in `dbc_genesis` had to be moved from `sn_domain` to the `client` module
in `safenode` because it uses the client for creating and working with the faucet.

The scope of various functions and modules were also changed from `pub(crate)` to `pub` so they
can be used by other crates.