chore: new version bumping script
Use the `release-plz` tool to bump the versions and generate changelogs for each crate.

Then use the git index to determine which crates had changes. This is a much simpler approach than
what we had while using `smart-release`.

Finally, a release commit is generated containing the crate names that were bumped. The release
process will use that information when adding assets to the Github release.