ci: release process
The release process now uses the `release-plz` tool.

The version bumping workflow uses a small script that runs `release-plz update` and parses its
output to determine which crates had version bumps. A commit is then generated based on that. Unlike
`smart-release`, `release-plz update` doesn't create a commit or tags. So now the version bumping is
only producing the commit and pushing that to `main`.

The release workflow is then kicked off, which builds the binaries, then uses the `release-plz
release` command to publish crates that were bumped using `update`. The same command also performs
Github Releases for the binary crates and creates tags for every crate that was published.

Finally, the binary assets are attached to the releases and uploaded to S3.