ci: workflow for building release artifacts
Provides a workflow that can be triggered manually for building release artifacts from either the
current `main` or a specified branch, for whatever reason you need at the time.

My particular use case is that I can use the artifacts to perform a 'manual' release run without
having a release commit, which can sometimes be useful for resolving situations in which releases or
crate publishing have went bad.