fix(replication): prevent dropped conns during replication
We were seeing a lot of OutboundFialure and closed connections leading
to dropped messages during data replication. This was turning up as
UnexpectedEof errors when a connection was closed midway through send
eg.

Seems like, due to _NOT_ responding to a request (using the protocol
wrongly perhaps?) we were causing awaiting connections to error out
as the receive stream was dropped unepectedly, causing a knock on to
anything else in flight on that same connection.