ci: upload release and s3 assets at the same time
This is to make sure that the versioned assets are not overwritten by another build.

Due to the way this process works, it's possible that all the binaries don't get updated for the
same release. The `upload-release-assets` target is checking which binaries are released and will
only upload the assets in the case of a release. The way it was setup previously, a new build would
have overwritten the versioned artifacts that were on S3 because it was just uploading them all.
Technically the build should be the same, but this just makes sure a versioned asset is not
overwritten.

We don't bother doing the same thing for the 'latest' release that goes to the S3 bucket, because
those are only used for internal purposes.