feat: use standarised directories for files/wallet commands
BREAKING CHANGE: use `~/.local/share/safe/wallet` as the root directory for wallet storage and
`~/.local/share/safe/client` for files related to downloads/uploads.

The previous directories were `~/.safe/wallet` and `~/.safe/client`, which we can phase out in
preference of using directories in the XDG standard.

Usage of a `RECEIVED_DBCS_PATH` environment variable was added for users who may feel the default
directory a bit unwieldy for working with the `wallet deposit` command, especially on Windows, where
the `AppData` directory is marked as hidden and so will only be shown by Windows Explorer if users
have viewing hidden files enabled.

Also provided and/or clarified documentation for other commands.