refactor: obtain genesis peer id directly
BREAKING CHANGE: Remove the `verify-nodes` feature in favour of just including this code in the
binary. It only increases the size by a few megabytes and the binary is much more useful with this
code.

Refactor the genesis launching process to use the gRPC generated code to obtain the genesis peer ID,
rather than using the `safe_rpc_client` example. The latter was actually built during the process,
which involved having the source code checked out. This is probably an assumption we shouldn't make
of a released binary. It's intended to be able to work stand alone, without the source.