feat: provide `--log-output-dest` arg for `safenode`
BREAKING CHANGE: the `--log-dir` argument has been removed in favour of `--log-output-dest`.

This argument uses a 'special' value, which can be "stdout", "default", or a custom directory path.
The "default" is the default root directory, plus "logs" appended to that path.

The `sn_logging` crate now provides a `LogOutputDest` enum which makes for a nicer interface on the
logging initialisation, removing the need for an `Option<PathBuf>` to specify output to a file.