feat: introduce `--log-format` arguments
BREAKING CHANGE: for both `safe` and `safenode` the `--json-log-output` argument is removed in
favour of this new `--log-format` argument.

The argument functions in a similar fashion to `--log-output-dest`, in that it uses a 'special'
value, which for the time being can either be "default" or "json". It has the potential to be
expanded.