ci: update node logging paths
In light of the changes to the logging setup, the testnet tool is modified such that it will use the
default root and logging directories.

The merge workflow is also updated to look for logs in the new directories. For the correct
usage of paths, in some cases this meant using different mechanisms on Windows vs. Unix.

The memcheck workflow was also updated to the same effect, but it only runs on Linux, so
cross-platform considerations weren't required.