feat: randomly select unused port
If a port is not specified, an unused port will be selected at random. If you pass a port with the
value 0 to the `libp2p` infrastructure, it will give you a port it says is provided by the OS, but
it doesn't mention whether it guarantees that the port is not in use. I was not able to determine if
this is the case.

We can add this simple mechanism to our own code to select a port that is unused. This will help for
running multiple nodes on the same machine, particularly in the context of the testnet tool.