ci: use custom action for testnet management
The Powershell script for collecting log artifacts was incorrect, and instead of fixing the
same script in multiple places, this prompted the creation of a custom action to start and stop the
local network. After the network is stopped the logs are tar'd and uploaded as artifacts.

There are three other workflows that launch a local testnet. They use the action to start the
network, but their logging process involves also collecting logs for heaptrack, so their logging
scripts remain directly in the workflows. We can make an extension to the custom action later for
supporting a heaptrack node.