ci: fix issues in dbc spend test and memcheck workflow
The `local-discovery` feature was unintentionally removed from the client and node binaries for the
DBC spend tests.

For the memcheck workflow, we had to prevent `SAFE_PEERS` from being reset in the new custom action
because the node that's runner with heaptrack is the only one that stays consistent through the
test.