chore(networking): only remove data as a last resort
We'd added constant pruning as records were PUT
(in order to tie storage price to data stored)
but this reduces the cache of data on the network.

Instead we can measure relevant data only as part of our calculation
for storage space, and retain our caching of data until we need that
space.