chore: store uploaded files list as text
For the uploaded files list, rather than serialize to binary, we use a simple text format. The user
can now see the addresses of all the files they've uploaded, and use this information to retrieve
them, if need be. We are also now just using one file, appending to it for each upload.

Now that this file is human readable, the `files ls` command was deemed unnecessary and was
therefore removed.

With respect to the `files` commands, I've also taken the opportunity to make the language in the
text output more concise and matter of fact. For example, we don't really need to use things like
exclamation marks in the text.