refactor: pay_for_chunks returns cost and new balance
Rather than the client printing out information relating to the cost of the chunks and the resulting
new balance, the information is returned to the caller, who can then print it out.

The error handling for storing the wallet is also changed to propagate the error rather than
suppressing it by printing output.