refactor: use one files api and clarify variable names
The `file_api` was being created multiple times when it didn't seem to need to be. It is easier to
pass around the `file_api` that you create once, rather than the client and the wallet directory
path.

I also renamed `root_dir` to `wallet_dir_path` because there didn't seem to be any need for it to
not be more specific as to what it actually refers to.