feat: use progress bars on `files upload`
Change the user interface for the `files upload` command to use progress bars to indicate the status
of the following operations:

* Splitting the input files into chunks
* Uploading the split chunks
* Verifying the chunks that were uploaded
* Re-uploading any chunks that failed verification

It uses a simple approach, which is to create a new `ProgressBar` for each of the operations, rather
than attempt to use a `MultiBar` or re-use the same progress bar. In this simple approach, the
`finish_and_clear` function removes the bar from stderr, then a new bar is created for the next
operation.

The `println!` output from the client was removed in favour of returning that information back to
the caller.

There is much less text in the output of the command now, but it hopefully communicates everything
that's important.