feat: provide `install` command
The install process is:

* Get the latest version number for `safenode`.
* Download a `safenode` archive for the current platform, using that version number.
* Extract it to a system wide location.
* Create a user for running the service if it doesn't already exist.
* Obtain a free port for the node service.
* Obtain a free port for the RPC service.
* Create a service for `safenode`, which will run the node and the RPC service.

This does *not* start any of the installed services--that will be a separate command.

The testing approach here is behaviour-driven unit tests, rather than integration tests, though I do
plan to add an integration test which will function as more of an end to end test that will grow
over time. The unit tests can be more numerous and cover small differences in configuration without
a full setup, which will require root access. The testing approach influenced the design here, with
the `install` function accepting traits which can be supplied mocks during the tests.

This commit also defines a merge workflow which runs the standard build and unit test jobs.