test: provide initial integration tests
What's provided in these tests is a first step, but the intention is for them to evolve into end to
end tests that will install, start/stop services, add new services, and so on. They will run by
executing the safenode-manager binary.

These tests are intended to demonstrate that we are running on our supported platforms without
introducing any major regressions. They will follow one config path without delving into different
configuration options--we'll leave that to the unit tests.

They need to run as the root user, and ideally, in a CI environment with a fresh build agent,
because they will create real services and not attempt to clean themselves up. Run on your own
machine at your own risk.