chore: introduce sn_rpc_client
Provides a client which is a wrapper around the Protobuf-generated code and types. The intention is
to avoid redefining that code in different places where the RPC protocol is used.

Right now it doesn't implement the full protocol; for example, the node events stream is missing.
However, it should serve as a starting point, and it is being used by the node manager.

The `sn_node_rpc_client` crate was extended so that it now doubles as a library and a binary, with
the client being added to the library. The binary was updated to use the client, where appropriate.