chore: specify root and log dirs at install time
To facilitate starting a node again with the same data, we need to supply the `--root-dir` argument
at install time. In addition to that, we need to actually create the directory and assign it the
correct permissions such that the user running the service will be able to write to it.

The log dir is also specified. This makes the logs available at system-wide locations rather than
the default directories, which is the data directory of the user running the service, which can make
them difficult to access, especially on Windows. Despite being at a system-wide location, on Linux,
the directory is assigned permissions to be written to by the `safe` user.

I tested the configurations on Windows and Linux.

I have also added a utility Vagrantfile here for quick testing. The reason for using a VM is because
you need to run commands as root and actually install services. I didn't want to have to continually
clean up that mess on my dev machine.