feat: provide `status` command
The command provides both a summary and a detailed view of installed services.

It checks to see whether the service process is still running, because it's possible the process
could have died since the last time we checked.

The data and log directories were added to the `InstalledNode` struct so that we could keep track of
these and display them in the detailed status view since it seems like useful information to have.