feat: provide `remove` command
Removes an added node, which will delete its data and log directories. The node is marked as
removed rather than actually being removed from the node registry just because the service number is
incremented by one each time a new node is added, so it might be a little odd if numbers were
missing in the sequence.