refactor: take either SAFE_PEERS or --peer
This is the conventional way to set arguments in clap. However, the
churn test launched a genesis node, but also a testnet that includes a
'new' genesis node. The testnet would pass --peer for the 'new' node,
while we'd actually want our first genesis node to be used to bootstrap,
which is set in SAFE_PEERS.