feat: provide `upgrade` command
The upgrade process simple:
* Get the latest version
* Download the new binary
* Stop the running node
* Overwrite the old binary with the new binary
* Start the node

It avoids doing anything if all the nodes are already at the latest version, and it also only
downloads the new binary once, even if there are multiple nodes to upgrade.

Also provided here is a Bash script that can spin up a quick test machine on Digital Ocean. This is
useful for adding nodes that connect to real networks without having to be concerned about NAT on
the local VM. I tried to use the DO provider for Vagrant, but sadly it didn't work correctly; I
think it was out of date with the current API.

The branch for the `remove` command was also rebased on top of this commit. As a result, the
`safenode_path` on the `Node` struct was also set to an optional field to make it consistent with
`data_dir_path` and `log_dir_path`.