feat: provide `--url` argument for `add` command
The URL enables use of a custom `safenode` binary that was built and uploaded somewhere. The URL
must point to a zip or gzipped tar archive which contains the binary.

The main reason for this is to support the creation of testnets that want to use a `safenode` binary
that is built from a fork of the main repository.