feat: custom port arguments for `add` command
When we're launching a genesis node, it needs to run on known ports, so we need to be able to
specify them.

It only really makes sense to use these arguments if a single service is being added, since multiple
services can't use the same port. So the code asserts that they can only be specified when one
service is being added.