chore: add data path field to node info
Since we provide the logs path, it's useful for the node manager to also have access here to the
data path, which it can provide to the user.

I also wanted to provide the port the node is running on, but it's not easy to get access to that if
the port is assigned by libp2p.