feat: provide `run` command
NOTE: this commit also contains the corresponding `kill` command. I accidentally amended it and
couldn't untangle the changes from it without losing them.

Spin up `safenode` processes, along with a faucet, to run a local network. These processes are not
defined as services because they are not intended to be long running; they are for experimentation.

As in the `testnet` binary, the network is validated, but unlike in `testnet` the logs are not
inspected. We just query each RPC service to check that peers are connected.

At present you need to supply paths to both binaries, but this will be extended shortly to provide
the ability to run downloaded binaries.

The node registry was extended to keep track of the faucet process so it can easily be cleaned up in
a subsequent `kill` command.

The `status` command is also extended to accommodate the local network.