feat: download binaries in absence of paths
If `--node-path` or `--faucet-path` arguments are not supplied, the latest versions of released
binaries will be downloaded. This enables `safenode-manager run` without any arguments. The ability
to get specific versions is available through `--node-version` and `--faucet-version` arguments.

The `kill` command was also improved to accommodate a scenario where the local network doesn't spin
up properly; you can use it to clean up an incomplete network.