feat: provide `--first` argument for `safenode`
In the previous setup, if the `network-contacts` feature was enabled, running `safenode` with no
`--peer` args caused the peers to be retrieved from the default network contacts file. In the
situation where we're launching a genesis node, we don't want this behaviour, and hence we introduce
a `--first` flag to make a distinction.

The `safenode` binary was changed to remove the use of `unwrap_or` when calling
`get_peers_from_args` so that an error will actually occur when peers are not obtainable, rather
than swallowing it by returning an empty peer list.

The `testnet` binary was also updated to use the `--first` argument when a new network was being
created. This necessitated a change to the memcheck workflow, which starts a testnet in a slightly
unusual fashion. The `--first` argument had to be applied to the initial node that is launched, and
then the subsequent local testnet launch had to be changed to use a join rather than starting a new
network. This is because the first peer launched by `testnet`, using `--first`, did not have the
initial node launched outside of `testnet` in its peer list. For the test to work correctly, an
environment variable was introduced to control the starting port deployment inventory, because using
a join network means the port range starts one digit higher. We will be able to remove this when we
switch over to using the node manager.

BREAKING CHANGE: the `parse_peer_args` function was renamed `get_peers_from_args` and the error
handling was changed. The new function name is semantically more accurate, and because
`sn_peers_acquisition` is a library crate, we should prefer an `Error` type rather than using
`eyre`.

Tried to add some unit tests for the `get_peers_from_args` function, but Tokio being an optional
dependency proved to be problematic.