chore: use `--first` arg for genesis node
A new `--first` argument has been added to `safenode` to facilitate the distinction between when we
launch a genesis node and when we want the default network contacts to be used.

We now update the local `run` command to use this argument to correctly launch a genesis node.

Also updated the year in the copyright notices on all the source files. This is something that's not
legally necessary, but it's the way the licensing verification tool works.