chore: tidy peer management for `join` command
The `peers` argument is no longer declared with an `Option` and the manual parsing of `SAFE_PEERS`
is removed. If the user doesn't specify any peers using `--peer` or `SAFE_PEERS`, the
`get_peers_from_args` function will return a particular type of error. In this case, we can just
assume the user wants the nodes to join an existing local network.