chore: additional info in `status` cmd
The binary path is provided because this is useful to know, particularly for nodes installed as
services.

The number of connected peers are also included, because again, this can be useful to know if the
network has initialised itself correctly, and how it might change over time.

There was a bit of a refactor in here to have the `run_node` function return a `Node` rather than
a multiaddr, because there's a function on the node now to get the multiaddr. The refactor also
eliminated a test case, which was no longer necessary.