chore: provide script for local network
The difference between this and the `run` command is that it sets up the nodes as services rather
than just running them as processes. It requires root to run.

The script will be used by the local testnet Github Action to start a network for the backwards
compatibility test. The network will then be upgraded.