ci: build the node manager on network start
When launching a local network, for now, we need to build the node manager rather than download the
latest version. Once the node manager has been added to the `safe_network` release process, we can
change this.

The `sn-local-testnet-action` has been updated to assume that we're building the node manager from
`safe_network`, as opposed to its own repository.