ci: add `safenode-manager` to the release process
The node manager is now added in to the release process. As with the other binaries, it will have a
Github release and will be uploaded to S3.

A few crates are configured to prevent Github releases, but they will still have their binaries
uploaded to S3. This is with the exception of `sn_testnet` which will not have its binaries uploaded
anywhere because we're going to phase this crate out, in favour of the node manager.