chore: improve error handling for `add` command
If there is an attempt to add many services, we will try adding all of them, even if there is an
error at some point when one service is being added. The results will then be summarised at the end.
If there were any failed services, we'll return an error, but the services that were added
successfully will continue to be usable.