chore: improve error handling for `start` command
As with the `add` command, the `start` command can apply to many services. It is also changed such
that it will attempt to start every service, even if there are others that don't start. If any fail,
the command will return an error, but the services that were started will remain usable.

This behaviour is useful for the testnet deployment process.