docs: change local network section into guide
I took a series of fairly disparate sections and merged them into a guide that sets up and uses a
local network.

I opted to remove the use of `--release` arguments on commands, since they make the commands more
verbose and the code takes longer to compile. Debug builds should be adequate for an experimental
local network.

Also tidied up the use of headers to make them consistent.