fix: node manager `status` permissions error
A user reported an issue in the `status` command resulting in an error. The `status` command
actually attempts to report on two different things: installed services and a local network. It
first tried to report on the status of services, and in the process, it attempts to create the
`/var/safenode-manager` directory if it does not exist, because that's where the node registry file
is stored. So, if you ran the command as a non-root user, it would result in a permissions error.
The error didn't occur on on my machine because it happened to be the case that I had already
created that directory.

I've now changed the command so that it will first query whether a local network exists, then return
if that is the case, without trying to query for installed services. In the case of no local
network, when a query is made for installed services, the command will run without root access.