chore: improvements from dev feedback
Reduce the start up time for the local network by removing the wait in the `get_available_port`
function. This function is now only being used to select ports for the RPC service, so it shouldn't
be subject to the same problems we had with the node when it switched over to Quic.

Also, the `--clean` argument on the run command has been extended to kill any existing local
network.