ci: temporary debugging output for release process
Something is causing the build agent to run out of space. I have a suspicion it's possible that it's
something to do with how `release-plz` determines changes in the crates, and it could be struggling
because the number of crates in our workspace has grown.

The first thing is to enable debugging output, because I seen some `debug!` usage in the
`release-plz` codebase. Hoping maybe that could provide some clues if it is `release-plz` that is
causing the disk usage.

If not, each step has output to display the disk usage, which hopefully may give us other clues.

This commit can be reverted when we have solved the issue.