ci: use a different disk for the release process
For reasons not yet determined, the publishing process undertaken by `release-plz` is using up 20GB
of disk space, which results in the root file system on the Github Actions build agent being filled
up during the process, since the disk only has 20GB of free space.

It turns out there's another disk at `/mnt`, which has over 60GB of space, so we should be able to
run the release process from there, until we can resolve the problem with `release-plz`.

I tested the manual checkout on the merge workflow just to make sure it worked ok.

This also upgrades the version of `release-plz` to 0.3.43, in the hopes that it maybe uses less disk
space. Some `df -h` commands are left in so that we can see if there is any improvement.