feat(cli): get-faucet has maid address option
This introduces a second way to get tokens from the faucet using the cli

`safe wallet get-faucet <url>`
Retains the existing fixed amount of tokens into the current wallet.

`safe wallet get-faucet <url> <maid_address> <signature>
Receives a transfer from the maid snapshot for the amount of maid owned
by that address. The signature is a bitcoin signature by the maid
address with the message as the hex encoded safe wallet address that
will receive the transfer.