ci: upload binaries to s3 before publish
We found CI node manager upgrade tests did not work correctly when a release was running at the same
time, which is a common scenario. The reason is because the node manager uses `crates.io` to get the
latest version of the node and then attempts to download binaries from S3 based on that version
number, but those binaries were not yet available on S3.

We now change the process to upload the binaries to S3 before the publishing phase.