feat: provide `faucet stop` command
As with the `faucet start` command, this provides the corresponding `stop` command. I revisited the
idea of sharing code between the node and faucet services, but again came to the conclusion that
there wasn't much value added for the complexity it introduced. The result is that we have code
between them that only has small differences, but I think this is OK for now.