refactor: uniform service management
Introduce a service manager to treat the node, faucet, and daemon services in a uniform way. The
services can be started, stopped or upgraded in a generic way, but the code for creating them
remains specific. There is probably potential for service creation to be made generic, but it's more
complex.

There are now `NodeService`, `FaucetService`, and `DaemonService` types. They all implement a
`ServiceActions` trait, allowing them to be treated generically by a `ServiceManager<T>`, where `T`
is constrained to be a `ServiceActions` type.

BREAKING CHANGE: the `RpcActions` and `RpcClient` were moved out of `sn_node_rpc_client` and into
`sn_service_management` to avoid issues with circular referencing.