chore: extend `status` cmd for faucet and daemon
This should help for using the faucet and daemon services in an idempotent way, while allowing for a
good user experience on the `add` command that will return an error if the services have already
been created.

The test util was also updated to deserialize to a utility status type, which makes it easier to
work with than raw JSON `Value` objects.