feat: provide `faucet upgrade` command
Uses the generic service manager to provide a `faucet upgrade` command, which will be used during
the general upgrade process to either upgrade or downgrade the faucet.

Some code was common between the node and faucet upgrades, so that was refactored to be shared
between the two.

Also clarifies some of the language used on the command documentation.