chore: run `safenodemand` service as root
Creating a user for running the `safenodemand` service doesn't make as much sense because the user
must have root access. We can leave it up to the user to create an account and set it up for
passwordless sudo, then provide an optional `--user` argument. For now we don't have this, and just
go with root.