feat: uniform behaviour for all `add` commands
The three `add` commands for `safenode`, `safenodemand` and `faucet` now behave in a uniform way,
all supporting `--path`, `--url` and `--version` arguments. These enable a binary to be provided via
a path, a URL pointing to a tar/zip containing the binary, or by a specific version number,
respectively.

The `safenodemand` service also now supports running as a non-root user.