feat: provide `--rpc-port` arg for `add` cmd
This feature was requested in community feedback. It behaves the same way as the other port
arguments, in that it can be a range.

Some of the `add_node` tests were changed to remove the use of a custom RPC address argument because
the test case did not concern itself with customising the RPC address.