feat: arguments can be used multiple times
The `--service-name` and `--peer-id` arguments are changed from single use to multiple. This feature
was requested in community feedback.

It was only requested for one command, but a few different commands use these arguments, so all were
updated to behave consistently.

The integration tests were updated to make use of the arguments multiple times, which will function
as a test for the change.