fix: permit removal of manually removed services
A community user manually removed one of his service file definitions and he then discovered that
the node manager would not allow the service to be removed from its registry.

These changes permit the manual removal of the service definition and also the data and log
directories. If they have been removed, we just proceed to remove the service from the node
manager's registry. This solution has only been tested for Systemd-based services. It would need to
be expanded and tested for other platforms and service managers.

It would have been possible to submit a PR for the service manager crate to behave in an idempotent
way and not error if the service file had been removed, but it seems correct for the service manager
to error in that case. We can handle the error in our application.