fix: preclude removed services from ops
The `start`, `stop`, `remove` and `upgrade` commands are changed such that services that have been
removed are not considered when operating over a list of services.

The `start` command in particular was subject to a bug where it tried to `start` removed services,
and it prevented it from being used to start all services with one command.