chore: refresh node registry before commands
The node registry is updated to the best of our ability before we run the `start`, `stop`, `remove`
or `upgrade` commands.

Here is an example of why this should apply. If we start a service, the node registry now stores the
PID for the process that is started by the service. It's possible for a user to kill the process
outside of both the node manager and the OS service infrastructure. However, our service definitions
have a restart policy, meaning when the user kills the process, the service infrastructure will
start another one. In this scenario, the node registry now has a PID for a dead process. So if we
then use the node manager to try and stop the service, it determines there is no process with the
dead PID, and declares the service has already been stopped.

The solution is to refresh the registry and get the most up-to-date status before running any of the
control commands.