ci: fix node manager binary packaging
In the packaging process, the crate name for the node manager was incorrectly renamed to
`sn-node-manager`. Confusingly, this is the correct name for the crate, but the directory and crate
name are mismatched. For clarity, I've changed the name of the variable to `crate_dir_name`.

Also, the `sn-node-manager` string was removed from the supported binary array, because that's the
name of the crate, not the binary.