refactor: provide `local` subcommand
A `local` subcommand is introduced for working with local networks.

We observed a user make the mistake of using the `run` command to try to connect to a remote
testnet, so it seems like a `local` subcommand would be useful for clarity.

There are no functional changes here, it's just some reorganisation.