chore: clarifications to version bumping script
There was an issue with the script always applying 0 for the pre-release identifier. We got into a
situation where it actually needed to be incremented by 1, because there was already a crate
published at the 0 version.

Added some text output to help clarify what the script is doing.

Also tidied the style by consistently applying Bash syntax (as opposed to POSIX compatible) and a
tab size of 2.