feat: distinguish failure to start during upgrade
It's possible for a service to be upgraded but then not subsequently start. In this case, it has
still been upgraded to a new version. It's worth making a distinction to the user between an actual
error in the upgrade process, or a failure to start, because in the latter case, they do actually
have an upgrade. They can then take action to try and start their services again.

As part of this change, the start process attempts to find whether the service process did indeed
start, because you don't always seem to get errors back from the service infrastructure. We also
make sure that we return an error if there was a failure with the upgrade process for any services.
This is necessary for visibility on our own deploy process.