feat: provide `--path` arg for `upgrade` cmd
This feature was requested in community feedback.

It allows users to build a custom binary and use that for the upgrade. We use the 'force' mechanism
here to cause the upgrade process to accept this binary, regardless of its version number. The user
may have built it from an experimental branch which could have a version number that's less than the
current version.

If the user is building their own binary from source, they are likely advanced, and aware of the
risks, so it seems reasonable to apply the force.