fix: do not delete custom bin on `add` cmd
When the `--path` argument is specified on the node manager's `add` command, a custom binary is
supplied, and now, the source of `--path` is not deleted.

By default, the `add` command downloads the latest version of a binary to a temporary location, and
that binary is then copied to the service location. Having finished with it, the node manager then
cleans up the temporary binary. When the `--path` argument is used, it goes through the same code
path, and hence the source of `--path` was being deleted as a side effect. Now, a flag is used to
indicate whether the binary should be deleted, and it is set to `false` when the `--path` argument
is supplied.

The `--path` argument is being used when users build their own `safenode`, so they are expecting
that the built binary will still exist.