feat: provide `autonomi-launcher` binary
This small binary attempts to detect the presence of a terminal executable, then uses it to run the
`node-launchpad` TUI inside it.

The TUI is downloaded and stored at `~/.local/bin/autonomi-launcher` on Linux/macOS and
`%USERPROFILE%\safe\autonomi-launcher.exe` on Windows. Right now, the version it downloads is hard
coded to `0.1.0-alpha.0`. This can be changed quite easily once we get a new release.

The launcher is added to the release process as a new binary. However, there is a complication not
yet dealt with. The `autonomi-launcher` and `node-launchpad` binaries are in the same crate, so the
process for uploading assets to Github releases will need to be modified to accommodate that. We can
come back to it.