feat: run safenode services in user mode
The `Systemd` and `Launchd` service managers support running services in user mode, and the
`service-manager-rs` crate already accommodates it, so we can support this scenario.

Services are managed in user mode depending on whether the node manager is running elevated or not.
In some cases, elevated privileges are still required, because not all service managers support
user-mode services. For example, Alpine uses OpenRC, which does not have support. A service running
in user-mode requires an active user session. This may not always be appropriate, particularly for
advanced users who want to leave nodes running permanently in the background.

The RPC daemon and the faucet services are still defined to run as system-wide services and they
don't support running in user-mode, because it doesn't really make sense to require an active user
session for these services.