chore: change terminal launch behaviour
On Linux, the `TERM` variable is not really appropriate for selecting which terminal is in use, so
this has been removed. It can be set to a completely different value, like `xterm-256color`, which
does not necessarily correspond to the running terminal. The `sudo` was also removed from the
terminal launch mechanism.

On the Windows front, I updated Windows Terminal to not use the `/c` argument. It simply just passes
the command to the terminal.

Also fixed another couple of issues:
* Compile the node manager correctly on Windows.
* The use of `<service>` in doc comments was being flagged as an unclosed HTML tag.