feat: provide `--upnp` flag for `add` command
The `--upnp` flag that is available on the node is now also available on the node manager `add`
command.

The flag only applies to a node that was built with the `upnp` feature. Rather than also using the
feature in the node manager, the command is documented to specify that it should only be used with
a `safenode` binary that has the feature. So for now, this option will probably be used in
conjunction with `--path`, to provide a node that has been built with `upnp` enabled.