fix: retain options on upgrade and prevent dup ports
The following options that can be specified when the service is created are now retained on an
upgrade:

* Custom `--node-port` value
* Custom `--metrics-port` value
* The `--home-network` flag
* The `--upnp` flag

Tests were added for all of these, and in addition, a test was added for the retention of custom RPC
ports, though these were already being retained on upgrade.

Although the node port was being tracked as part of the `listen_addr` field in `NodeServiceData`, a
new `node_port` field was explicitly added, because it's possible that a service can be upgraded
before it starts, and the `listen_addr` is not assigned until the service starts.

I also discovered the `add` command did not prevent specifying custom ports that were already in
use. That was fixed too, and some tests were added.