chore: reconfigure local network owner args
The `local run` and `local join` commands had an optional `--owner` argument being used in testing
scenarios to pass `--owner` to each `safenode` process. In its absence, the application code was
checking if it was running in a CI context, and if so, it was assigning an individual owner to each
node. This owner would be in the form "node_X", where "X" was replaced with the node number. If the
code determined it wasn't running in CI, it would still launch the node with `--owner`, assigning a
default value.

This behaviour has been changed such that `safenode` will only be passed `--owner` if it has also
been used on the node manager `local run/join` commands. This lets `safenode` itself provide a
default value. Now, the `local run --owner` argument assigns the same owner to each `safenode`
process in the local network, and a new argument, `--owner-prefix`, assigns an individual owner to
each node, based on a prefix. For example, if you use `--owner-prefix node`, the nodes will be
assigned owners "node_1", "node_2", and so on.

We need to configure the CI system to specify owners where it needs to. This new setup prevents CI
details leaking in to the application code.